requirements.txt
flask
np
numpy
Pillow
